refactor(processing): rename global DEM module
